codetomd: convert shell to markdown
A simple python script to convert a shell script to a Markdown file with code sections

Write your shell script with comments, then convert it to a "tutorialesque" document

Only top-level comments (ones whose '#' is the first token on the line) will be treated as documentation.

Anything indented, or not a comment, will be encapsulated as a code block.